fix(release): make runtime-deps bare mode marker-aware (#661)
Bare `release runtime-deps` (no --staging) now strips `# gaia:maintainer-only` blocks from each .sh source before extracting path refs, so it agrees with the authoritative post-scrub --staging run instead of false-positiving on maintainer-only references. The strip primitive (stripMarkerBlocks) is extracted from scrub.ts into a shared marker-strip module so there is one marker parser, not two.

Closes #661
